+++ pyproject.toml
@@ -80,7 +80,7 @@ include = ["bitsandbytes*"]
 version = {attr = "bitsandbytes.__version__"}
 
 [tool.pytest.ini_options]
-addopts = "-rP"
+addopts = "-rP -m 'not slow and not benchmark and not deprecated'"
 #    ; --cov=bitsandbytes
 #    ; # contexts: record which test ran which line; can be seen in html coverage report
 #    ; --cov-context=test
